Refine artist filters layout and extract components.
Compact filter selects to align with social icons, rename labels, reorder Zora, and split ProfileStats/FilterSelect/ArtistPageContent plus shared types.
